Python Automation API task
This tasks features a HTTP request methods namely, GET, POST, PUT and DELETE. Json and Requests were imported to execute the python codes. HTTP is a means of comunication by web API'S. HTTP protocol follows a client server model where clients communicate with server by sending request and waiting for responses. GET request method simply requests for information to retrieve data and should not cause any data modificaton. You can retrieve employees, students etc data via GET requests. POST are used to modify or cry a new resource. It means that whenever new student or employee is hired, a POST request is used to add the employee or student to the system. PUT request method is used to modify or update an existing data instead of creating a new one. So if employees  job title or department has changed PUT will be send to make the changes. DELETE request is used to remove data from the database. So if an employee leaves a company, the delete request method will be used. This project demonstrates how the HTTP methods are used in reality. This time in Python